pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "hipaa-mcp"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "fastmcp>=0.1.0",
    "chromadb>=0.5.0",
    "rank-bm25>=0.2.2",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "lxml>=5.0",
    "platformdirs>=4.0",
    "click>=8.1",
    "spacy>=3.7",
]

[project.scripts]
hipaa-mcp = "hipaa_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/hipaa_mcp"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "RUF"]

[tool.mypy]
strict = true
python_version = "3.12"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
